Source-Codes/server/scripts/Helpers/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blis==0.7.11
breadability==0.1.20
catalogue==2.0.10
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
DateTime==5.4
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
fastapi==0.110.0
filelock==3.13.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
lxml==5.1.0
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.14.3
packaging==24.0
pillow==10.2.0
preshed==3.0.9
pycountry==23.12.11
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
setuptools==69.2.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
sumy==0.11.0
sympy==1.12
textblob==0.18.0.post0
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.39.1
typer==0.9.4
urllib3==2.2.1
uvicorn==0.29.0
vaderSentiment==3.3.2
wasabi==1.1.2
weasel==0.3.4
wrapt==1.16.0
zope.interface==6.2
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
joblib==1.3.2
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
protobuf==4.25.3
psx==0.0.1
psx-data-reader==0.0.6
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.13.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
